Updating

+++ src/EventStore.Core/Data/EventNumber.cs
@@ -1,8 +1,6 @@
-namespace EventStore.Core.Data
-{
-    public static class EventNumber
-    {
-        public const int DeletedStream = int.MaxValue;
-        public const int Invalid = int.MinValue;
-    }
+namespace EventStore.Core.Data {
+	public static class EventNumber {
+		public const long DeletedStream = long.MaxValue;
+		public const long Invalid = int.MinValue;
+	}
 }
